% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{anpl}
\alias{anpl}
\title{Adaptive non-parametric learning function}
\usage{
anpl(dataset, concentration, n_bootstrap = 100,
  posterior_sample = NULL, gamma_mean = NULL, gamma_vcov = NULL,
  threshold = 1e-08)
}
\arguments{
\item{dataset}{The dataset with classes and features (in a specific format
that will be changed)}

\item{concentration}{The parameter \code{c} in the paper (page 3, formula 3),}

\item{n_bootstrap}{The number of bootstrap samples required.}

\item{posterior_sample}{The function can take samples from the posterior to
generate non-parametric-learning samples. Or it can take NULL and the
posterior is assumed normal N(\code{gamma_mean}, \code{gamma_vcov}).}

\item{gamma_mean}{In case \code{posterior_sample} is NULL, the mean for the
centering model (equation 9, page 4)}

\item{gamma_vcov}{In case \code{posterior_sample} is NULL, the
variance-covariance of the centering model for gamma (equation 9, page 4).}

\item{threshold}{The threshold of stick remaining below which the function
stops looking for more stick-breaks. It correspondes to epsilon in the
paper, at the bottom of page 5 and in algorithm 2 in page 12.}
}
\value{
A matrix of bootstrap samples for the parameter of interest
}
\description{
\code{anpl} returns samples of the parameter of interest
}
\details{
This function implements the non-parametric-learning algorithm, which is
algorithm 2 in page 12 in the paper. It uses a mixture of Dirichlet processes
and stick-breaking to find the number of posterior samples and logistic
regression to find the randomized parameter of interest. For examples, see
the vignette.
}
