% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{get_file}
\alias{get_file}
\title{Get a file from extdata by name}
\usage{
get_file(name)
}
\arguments{
\item{name}{The filename that is requested}
}
\value{
The requested file
}
\description{
Get a file from extdata by name
}
\examples{
f <- get_file('bayes_logit.stan')
writeLines(readLines(f))

}
