% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{get_german_credit_dataset}
\alias{get_german_credit_dataset}
\title{Function to load the dataset that ships with the package.}
\usage{
get_german_credit_dataset(scale = TRUE)
}
\arguments{
\item{scale}{Whether to scale the features to have mean 0 and variance 1.}
}
\value{
A dataset in the right format
}
\description{
Function to load the dataset that ships with the package.
}
\examples{
german <- get_german_credit_dataset()
head(german$y)
head(german$x)

}
