% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{get_german_credit_file}
\alias{get_german_credit_file}
\title{Get the file with the German Statlog credit dataset}
\usage{
get_german_credit_file()
}
\value{
A file with the plain-text raw data for the German Statlog credit
  that ships with this package (extension \code{.dat}).
}
\description{
Get the file with the German Statlog credit dataset
}
\examples{
f <- get_german_credit_file()
\dontrun{
file.show(f)
}

}
