% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{get_rstan_file}
\alias{get_rstan_file}
\title{Get the RStan file with Variational Bayes for comparison}
\usage{
get_rstan_file()
}
\value{
An RStan file with the model for variational Bayes that ships with
  this package (extension \code{.stan}).
}
\description{
Get the RStan file with Variational Bayes for comparison
}
\examples{
f <- get_rstan_file()
\dontrun{
file.show(f)
}

}
