% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\docType{package}
\name{PosteriorBootstrap}
\alias{PosteriorBootstrap}
\alias{PosteriorBootstrap-package}
\title{The 'PosteriorBootstrap' package.}
\description{
What the Package Does (One Line, Title Case)
}
