% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{run_variational_bayes}
\alias{run_variational_bayes}
\title{Run variational Bayes}
\usage{
run_variational_bayes(x, y, output_samples, beta_sd,
  stan_file = get_stan_file(), iter = 10000, seed = 123,
  verbose = FALSE)
}
\arguments{
\item{x}{The features of the data for the model}

\item{y}{The outcomes of the data for the model}

\item{output_samples}{The number of output samples to draw}

\item{beta_sd}{The standard deviation of the prior on the parameters}

\item{stan_file}{A custom Stan file, if different from the default which
ships with the package and models Bayesian logistic regression}

\item{iter}{The maximum number of iterations for Rstan to run}

\item{seed}{A seed to start the Rstan sampling}

\item{verbose}{Whether to print output from RStan}
}
\value{
Matrix of size `output_samples`x`ncol(x)` drawn from the RStan
  model applied to `y` and `x`
}
\description{
\code{run_variational_bayes} returns samples of the parameters estimated with
`rstan` on the data provided in the arguments.
}
