% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{script}
\alias{script}
\title{Wrapper function for the script part of the code.}
\usage{
script(use_bayes_logit = TRUE, verbose = TRUE)
}
\arguments{
\item{use_bayes_logit}{Whether to use this package or the alternative
from Kaspar Martens}

\item{verbose}{Whether to print statements}
}
\description{
Note: this function takes several hours to run on a mac laptop.
}
