% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{stick_breaking}
\alias{stick_breaking}
\title{Stick-breaking depending on a concentration parameter.}
\usage{
stick_breaking(concentration = 1, min_stick_breaks = 100,
  threshold = 10^(-8))
}
\arguments{
\item{concentration}{The parameter \code{c} in the paper (page 3, formula 3),
which is an effective sample size.}

\item{min_stick_breaks}{The minimal number of stick-breaks.}

\item{threshold}{The threshold of stick remaining below which the function
stops looking for more stick-breaks. It correspondes to epsilon in the
paper, at the bottom of page 5 and in algorithm 2 in page 12.}
}
\value{
A vector of stick-breaks summing to one.
}
\description{
\code{stick_breaking} returns a vector with the breaks of a stick of length 1
}
\details{
This function implements the stick-breaking process for non-parametric
learning described in section 2 of the supplementary material. The name
"stick-breaking" comes from a stick of unit length that we need to break into
a number of items. This code implements algorithm 2 and the stick-breaking
function calculates the parameter T in algorithm 1, which is the only
difference between the two algorithms. The code uses the Beta distribution as
that distribution is part of the definition of the stick-breaking process.
}
\examples{
stick_breaking(1)
stick_breaking(1, min_stick_breaks = 10)
stick_breaking(1, min_stick_breaks = 10, threshold = 10 ^ (-8)

}
