% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorBootstrap.R
\name{get_stan_file}
\alias{get_stan_file}
\title{Stan file with Bayesian Logistic Regression.}
\usage{
get_stan_file()
}
\value{
An RStan file with the model for variational Bayes that ships with
  this package (extension \code{.stan}).
}
\description{
Stan file with Bayesian Logistic Regression.
}
\examples{
f <- get_stan_file()
\dontrun{
file.show(f)
}

}
