% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcb_vb_classn_sb.R
\name{load_dataset}
\alias{load_dataset}
\title{Function to load dataset. Won't be necessary in the package}
\usage{
load_dataset()
}
\value{
A dataset in the right format
}
\description{
Function to load dataset. Won't be necessary in the package
}
