% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosteriorBootstrap.R
\docType{package}
\name{PosteriorBootstrap-package}
\alias{PosteriorBootstrap}
\alias{PosteriorBootstrap-package}
\title{A package with a parallel approach for adaptive non-parametric learning}
\description{
The PosteriorBootstrap package provides two categories of functions.
The first category returns or loads the system files that ship with the
package: get_stan_file, get_german_credit_file, get_german_credit_dataset.
The second category performs statistical sampling: draw_stick_breaks and
draw_logit_samples (for adaptive non-parametric learning of the logistic
regression model).
}
\details{
Please see the vignette for sample usage and performance metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alan-turing-institute/PosteriorBootstrap/}
  \item Report bugs at \url{https://github.com/alan-turing-institute/PosteriorBootstrap/issues}
}

}
\author{
\strong{Maintainer}: James Robinson \email{james.em.robinson@gmail.com}

Authors:
\itemize{
  \item Simon Lyddon \email{simon.lyddon@stats.ox.ac.uk}
  \item Miguel Morin \email{info@turing.ac.uk}
}

Other contributors:
\itemize{
  \item The Alan Turing Institute \email{info@turing.ac.uk} [copyright holder]
}

}
